Fix reference processing
* Part is never used as a portion of reference
* Valid reference portions are following ONLY:
  * For `list`, `img`, and `table`:
     `{label}` or `{chapter|label}`
  * For `fn`:
     `{label}`
  * For `chap`, `title`, and `chapref`:
     `{chapter}`
  * For `hd`:
    `{label}`, `{chapter|label}`, `{caption1|caption2|...|captionN}`,
    or `{chapter|caption1|caption2|...|captionN}`

This commit removes part support from Reference and Symbol,
and then adds strict validation for above patterns.
However, it is too hard to implement caption-path support
because it requires drastic modification of data models and their usage.
So, this just add friendly error message which leads to use of
label instead of caption-path.